Move `compilerV2` to `compiler`
