+++ inlang/source-code/paraglide/paraglide-js/src/compiler/aliases.ts
