+++ inlang/source-code/paraglide/paraglide-js/src/compiler/bundleIndex.ts
