+++ inlang/source-code/paraglide/paraglide-js/src/compiler/compileMessage.test.ts
