+++ inlang/source-code/paraglide/paraglide-js/src/compiler/compilePattern.test.ts
