+++ inlang/source-code/paraglide/paraglide-js/src/compiler/compilePattern.ts
