+++ inlang/source-code/paraglide/paraglide-js/src/compiler/registry.ts
