+++ inlang/source-code/paraglide/paraglide-js/src/compiler/runtime.ts
